Download youtubedown.js from bucket
