+++ lib/ntpUtil.js
@@ -1,0 +1,34 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+const childProcess = require('child_process')
+const fs = require('fs')
+const util = require('../lib/util')
+
+const generatePublicKeyAndID = (privateKeyFile) => {
+  childProcess.execSync(`openssl rsa -in ${privateKeyFile} -pubout -out public.pub`)
+  try {
+    // read contents of the file
+    const data = fs.readFileSync('public.pub', 'UTF-8')
+
+    // split the contents by new line
+    const lines = data.split(/\r?\n/)
+    let pubKeyString = ''
+    lines.forEach((line) => {
+      if (!line.includes('-----')) {
+        pubKeyString += line
+      }
+    })
+    console.log(`publicKey: ${pubKeyString}`)
+    const id = util.getIDFromBase64PublicKey(pubKeyString)
+    console.log(`componentID: ${id}`)
+    return [pubKeyString, id]
+  } catch (err) {
+    console.error(err)
+  }
+}
+
+module.exports = {
+  generatePublicKeyAndID
+}
