+++ package.json
@@ -69,6 +69,7 @@
     "upload-ntp-sponsored-images-components": "node ./scripts/uploadComponent --type ntp-sponsored-images",
     "upload-ntp-super-referrer-component": "node ./scripts/uploadComponent --type ntp-super-referrer",
     "upload-ntp-super-referrer-component-mapping-table": "node ./scripts/uploadComponent --type ntp-super-referrer-mapping-table",
+    "upload-youtubedown-component": "node ./scripts/uploadComponent --type playlist-youtubedown",
     "upload-user-model-installer-updates": "node ./scripts/uploadComponent --type user-model-installer",
     "upload-speedreader": "node ./scripts/uploadComponent --type speedreader-updater"
   },
